⚡️ Switch from `serde` to `serde_core`
This new split of `serde` crate, enables faster compilation by allowing
non-derive needing parts (such as this library) to be compiled in
parallel to `serde_derive` crate.

This also implies removal of `serde_derive` as an indirect dependency of
this crate.

A run of `cargo build --features serde` in this repo itself goes from 5s
to 1s (on my machine), which is a pretty big improvement.